backend/requirements.txt
fastapi
uvicorn
python-dotenv
gitpython
radon
cloc
requests
